structure streamlit skeleton with sidebar, main and documentation
