+++ application/functions.py
@@ -1,0 +1,6 @@
+import streamlit as st
+
+@st.cache
+def convert_df(df):
+    # IMPORTANT: Cache the conversion to prevent computation on every rerun
+    return df.to_csv().encode('utf-8')
